Extract throwAndExit
