+++ packages/astro/src/cli/index.ts
@@ -2,13 +2,11 @@
 import * as colors from 'kleur/colors';
 import yargs from 'yargs-parser';
 import { ASTRO_VERSION } from '../core/constants.js';
-import { collectErrorMetadata } from '../core/errors/dev/index.js';
-import { createSafeError } from '../core/errors/index.js';
-import { debug, type LogOptions } from '../core/logger/core.js';
+import type { LogOptions } from '../core/logger/core.js';
 import { enableVerboseLogging, nodeLogDestination } from '../core/logger/node.js';
-import { formatErrorMessage, printHelp } from '../core/messages.js';
+import { printHelp } from '../core/messages.js';
 import * as event from '../events/index.js';
-import { eventError, telemetry } from '../events/index.js';
+import { telemetry } from '../events/index.js';
 
 type Arguments = yargs.Arguments;
 type CLICommand =
@@ -201,29 +199,7 @@ export async function cli(args: string[]) {
 	try {
 		await runCommand(cmd, flags);
 	} catch (err) {
+		const { throwAndExit } = await import('./throw-and-exit.js');
 		await throwAndExit(cmd, err);
 	}
-}
-
-/** Display error and exit */
-async function throwAndExit(cmd: string, err: unknown) {
-	let telemetryPromise: Promise<any>;
-	let errorMessage: string;
-	function exitWithErrorMessage() {
-		console.error(errorMessage);
-		process.exit(1);
-	}
-
-	const errorWithMetadata = collectErrorMetadata(createSafeError(err));
-	telemetryPromise = telemetry.record(eventError({ cmd, err: errorWithMetadata, isFatal: true }));
-	errorMessage = formatErrorMessage(errorWithMetadata);
-
-	// Timeout the error reporter (very short) because the user is waiting.
-	// NOTE(fks): It is better that we miss some events vs. holding too long.
-	// TODO(fks): Investigate using an AbortController once we drop Node v14.
-	setTimeout(exitWithErrorMessage, 400);
-	// Wait for the telemetry event to send, then exit. Ignore any error.
-	await telemetryPromise
-		.catch((err2) => debug('telemetry', `record() error: ${err2.message}`))
-		.then(exitWithErrorMessage);
 }
